refactor(text): move tests and stories files to a separate folder
