+++ src/components/atoms/Text/__stories__/text.stories.tsx
@@ -1,12 +1,12 @@
 import { Meta, Story } from '@storybook/react/types-6-0'
 
-import { Text } from './text'
+import { Text } from '@/components/atoms/Text'
 
 export default {
   title: 'Text',
   component: Text
 } as Meta
 
-export const Basic: Story = (args) => (
+export const Default: Story = (args) => (
   <Text {...args}>Basic Text Component</Text>
 )
